+++ src/lib.rs
@@ -1,6 +1,7 @@
 #![feature(const_impl_trait)]
 #![feature(const_fn_trait_bound)]
 pub mod parser;
+pub mod weekday;
 
 // https://github.com/Geal/nom/blob/master/doc/choosing_a_combinator.md
 // https://stackoverflow.com/questions/56846090/nom-5-creating-a-combinator-using-another-parser-multiple-times
