+++ src/parser/mod.rs
@@ -1,3 +1,5 @@
+pub mod enum_utils;
+
 use std::fmt;
 use std::{
     num::{NonZeroI16, NonZeroI8, NonZeroU64, NonZeroU8},
@@ -348,26 +350,6 @@ impl fmt::Display for RecurRule {
     }
 }
 
-// TODO FIXME use chrono Weekday when arbitrary is implemented
-#[derive(PartialEq, Eq, Copy, Clone, Debug)]
-#[cfg_attr(feature = "arbitrary", derive(arbitrary::Arbitrary))]
-pub enum Weekday {
-    /// Monday.
-    Mon = 0,
-    /// Tuesday.
-    Tue = 1,
-    /// Wednesday.
-    Wed = 2,
-    /// Thursday.
-    Thu = 3,
-    /// Friday.
-    Fri = 4,
-    /// Saturday.
-    Sat = 5,
-    /// Sunday.
-    Sun = 6,
-}
-
 impl Default for RecurRule {
     fn default() -> Self {
         RecurRule {
@@ -482,16 +464,6 @@ impl fmt::Display for Frequency {
     }
 }
 
-pub const fn enum_element<'a, T: Copy>(
-    string: &'static str,
-    enum_element: T,
-) -> impl FnMut(&'a str) -> IResult<&'a str, T> {
-    move |input| {
-        let (input, _) = tag(string)(input)?;
-        Ok((input, enum_element))
-    }
-}
-
 pub fn freq(input: &str) -> IResult<&str, Frequency> {
     alt((
         enum_element("SECONDLY", Frequency::Secondly),
@@ -501,36 +473,6 @@ pub fn freq(input: &str) -> IResult<&str, Frequency> {
         enum_element("WEEKLY", Frequency::Weekly),
         enum_element("MONTHLY", Frequency::Monthly),
         enum_element("YEARLY", Frequency::Yearly),
-    ))(input)
-}
-
-impl fmt::Display for Weekday {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(
-            f,
-            "{}",
-            match self {
-                Weekday::Mon => "MO",
-                Weekday::Tue => "TU",
-                Weekday::Wed => "WE",
-                Weekday::Thu => "TH",
-                Weekday::Fri => "FR",
-                Weekday::Sat => "SA",
-                Weekday::Sun => "SU",
-            }
-        )
-    }
-}
-
-pub fn weekday(input: &str) -> IResult<&str, Weekday> {
-    alt((
-        enum_element("MO", Weekday::Mon),
-        enum_element("TU", Weekday::Tue),
-        enum_element("WE", Weekday::Wed),
-        enum_element("TH", Weekday::Thu),
-        enum_element("FR", Weekday::Fri),
-        enum_element("SA", Weekday::Sat),
-        enum_element("SU", Weekday::Sun),
     ))(input)
 }
 
@@ -564,6 +506,10 @@ pub enum RRuleDateOrDateTime {
 
 #[cfg(feature = "arbitrary")]
 use arbitrary::{Arbitrary, Result, Unstructured};
+
+use crate::weekday::{weekday, Weekday};
+
+use self::enum_utils::enum_element;
 
 #[cfg(feature = "arbitrary")]
 impl<'a> Arbitrary<'a> for RRuleDateOrDateTime {
